pyproject.toml
[tool.poetry]
name = "fastapi-ollama-langchain"
version = "0.1.0"
description = "FastAPI LLM model using langchain and Ollama"
authors = ["Ernesto Crespo <ecrespo@protonmail.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
langchain = "^0.1.12"
langchain-community = "^0.0.28"
langchainhub = "^0.1.15"
ollama = "^0.1.7"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
